src/main/java/com/akay/testproject/Service/CityService.java
package com.akay.testproject.Service;

import com.akay.testproject.Entity.City;
import com.akay.testproject.Repository.CityRepository;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import java.util.Optional;

@Service
public class CityService {

    private final CityRepository cityRepository;

    public CityService(CityRepository cityRepository) {
        this.cityRepository = cityRepository;
    }

    public City createCity(City city) {
        return cityRepository.save(city);
    }

    public City updateCity(Long id, City cityDetails) {
        Optional<City> optionalCity = cityRepository.findById(id);
        if (optionalCity.isPresent()) {
            City city = optionalCity.get();
            city.setName(cityDetails.getName());
            city.setCountry(cityDetails.getCountry());
            return cityRepository.save(city);
        }
        throw new RuntimeException("City not found with id " + id);
    }

    public void deleteCity(Long id) {
        cityRepository.deleteById(id);
    }

    public Page<City> findAll(Pageable pageable) {
        return cityRepository.findAll(pageable);
    }
}